chore(main): extract startup/lifecycle cluster into app-lifecycle (arch R6)
Pure move of the post-registerIpc() startup/lifecycle tail out of main.ts
into a new app-lifecycle.ts exporting wireAppLifecycle(deps): the
app.whenReady() startup flow (dock icon, visual-smoke seeding, credential
startup, window creation, background startup), runCredentialStartup /
runBackgroundStartup / ensureBootstrapConnection /
recoverInterruptedSessionsOnStartup, the window-all-closed and before-quit
handlers, and runBeforeQuitCleanup.

Startup ORDER and teardown ORDER are the product, so bodies are byte-identical
to their originals; every process-scoped collaborator is injected. The
single-instance lock and registerIpc() anchor stay in main.ts. The module-let
seams follow the R5 accessor precedent: setLookupPricing reassigns the pricing
lookup (read live by the session streamer + usage IPC) and getSettingsIpc reads
the handle assigned inside registerIpc(); configWatcher + beforeQuitCleanup
flags become app-lifecycle closure state. main.ts 1372 -> 1175 (-197).
